+++ config.py
@@ -1,10 +1,10 @@
 # Single-sourced project configuration values:
 
 # the full version, including alpha/beta/rc tags
-release = '0.8.2'
+release = '0.9.0'
 
 # the short X.Y version
-version = '0.8'
+version = '0.9'
 
 # distribution package name
 name = 'opyprint'