+++ tests/ppcontext/test_h_style.py
@@ -7,7 +7,7 @@ class Foo:
     def __str__(self, ppc: PPContext = None):
         ppc = ppc or PPContext()
         ppc("A Foo with:")
-        ppc({'k_1': "v_1", 'k_2': "v_2"}, style='yellow')
+        ppc({'k_1': "v_1", 'k_2': "v_2"}, style='cyan')
         return ppc.flush()
 
 
@@ -41,28 +41,32 @@ def test_ppc_style_2_misc():
 
     # print("#2.1:")
     result = ppc.format(123, style=('red', 'italic'))
-    # print(f"{result}\n{[result]}")
+    # print(f"\n{result}\n\n{[result]}")
     assert result == "\x1b[31;3m123\x1b[0m"
 
     # print("#2.2:")
     result = ppc.format(['abc', 123], style=('red', 'italic'))
-    # print(f"{result}\n{[result]}")
+    # print(f"\n{result}\n\n{[result]}")
     assert result == "\x1b[31;3m[abc, 123]\x1b[0m"
 
     # print("#2.3:")
     result = ppc.format(['abc', 'def', 'ghi', 'jkl'], style=('red', 'italic'))
-    # print(f"{result}\n{[result]}")
-    assert result == ("- \x1b[31;3mabc\x1b[0m\n"
-                      "- \x1b[31;3mdef\x1b[0m\n"
-                      "- \x1b[31;3mghi\x1b[0m\n"
-                      "- \x1b[31;3mjkl\x1b[0m")
+    # print(f"\n{result}\n\n{[result]}")
+    assert result == (
+        "- \x1b[31;3mabc\x1b[0m\n"
+        "- \x1b[31;3mdef\x1b[0m\n"
+        "- \x1b[31;3mghi\x1b[0m\n"
+        "- \x1b[31;3mjkl\x1b[0m"
+    )
 
     # print("#2.4:")
     result = ppc.format(Foo(), style=('red', 'italic'))
-    # print(f"{result}\n{[result]}")
-    assert result == ("\x1b[31;3mA Foo with:\x1b[0m\n"
-                      "\x1b[31;3m- \x1b[33mk_1: v_1\x1b[0m\x1b[0m\n"
-                      "\x1b[31;3m- \x1b[33mk_2: v_2\x1b[0m\x1b[0m")
+    # print(f"\n{result}\n\n{[result]}")
+    assert result == (
+        "\x1b[31;3mA Foo with:\x1b[0m\n"
+        "\x1b[31;3m\x1b[36m- k_1:\x1b[0m \x1b[36mv_1\x1b[0m\x1b[0m\n"
+        "\x1b[31;3m\x1b[36m- k_2:\x1b[0m \x1b[36mv_2\x1b[0m\x1b[0m"
+    )
 
 
 def test_ppc_style_3_dict():
@@ -70,10 +74,12 @@ def test_ppc_style_3_dict():
 
     # print("#3.1 - Case KVP-2:")
     obj = {'ef': 'Lorem ipsum dolor sit', }
-    result = ppc.format(obj, style=('red', 'italic'))
-    # print(f"{result}\n{[result]}")
-    assert result == ("\x1b[31;3mef: Lorem ipsum\x1b[0m\n"
-                      "  \x1b[31;3mdolor sit\x1b[0m")
+    result = ppc.format(obj, style=('white', 'italic'), key_style='red')
+    # print(f"\n{result}\n\n{[result]}")
+    assert result == (
+        "\x1b[31mef:\x1b[0m\x1b[37;3m Lorem ipsum\x1b[0m\n"
+        "  \x1b[37;3mdolor sit\x1b[0m"
+    )
 
     # print("#3.2 - Case KVP-1, KVP-3, KVP-2:")
     obj = {
@@ -81,13 +87,15 @@ def test_ppc_style_3_dict():
         'cd': 34,
         'ef': 'Lorem ipsum dolor sit',
     }
-    result = ppc.format(obj, style=('red', 'italic'))
-    # print(f"{result}\n{[result]}")
-    assert result == ("- \x1b[31;3mab: xy\x1b[0m\n"
-                      "- \x1b[31;3mcd:\x1b[0m \x1b[31;3m34\x1b[0m\n"
-                      "- \x1b[31;3mef: Lorem\x1b[0m\n"
-                      "    \x1b[31;3mipsum\x1b[0m\n"
-                      "    \x1b[31;3mdolor sit\x1b[0m")
+    result = ppc.format(obj, style=('white', 'italic'), key_style='red')
+    # print(f"\n{result}\n\n{[result]}")
+    assert result == (
+        "\x1b[31m- ab:\x1b[0m \x1b[37;3mxy\x1b[0m\n"
+        "\x1b[31m- cd:\x1b[0m \x1b[37;3m34\x1b[0m\n"
+        "\x1b[31m- ef:\x1b[0m\x1b[37;3m Lorem\x1b[0m\n"
+        "    \x1b[37;3mipsum\x1b[0m\n"
+        "    \x1b[37;3mdolor sit\x1b[0m"
+    )
 
     # print("#3.3 - Case KVP-1, KVP-3, KVP-5:")
     obj = {
@@ -95,33 +103,39 @@ def test_ppc_style_3_dict():
         'cd': ['abc', 123],
         'ef': ['abc', 'def', 'ghi', 'jkl'],
     }
-    result = ppc.format(obj, style=('red', 'italic'))
-    # print(f"{result}\n{[result]}")
-    assert result == ("- \x1b[31;3mab: xy\x1b[0m\n"
-                      "- \x1b[31;3mcd:\x1b[0m \x1b[31;3m[abc, 123]\x1b[0m\n"
-                      "- \x1b[31;3mef:\x1b[0m\n"
-                      "    - \x1b[31;3mabc\x1b[0m\n"
-                      "    - \x1b[31;3mdef\x1b[0m\n"
-                      "    - \x1b[31;3mghi\x1b[0m\n"
-                      "    - \x1b[31;3mjkl\x1b[0m")
+    result = ppc.format(obj, style=('white', 'italic'), key_style='red')
+    # print(f"\n{result}\n\n{[result]}")
+    assert result == (
+        "\x1b[31m- ab:\x1b[0m \x1b[37;3mxy\x1b[0m\n"
+        "\x1b[31m- cd:\x1b[0m \x1b[37;3m[abc, 123]\x1b[0m\n"
+        "\x1b[31m- ef:\x1b[0m\n"
+        "    - \x1b[37;3mabc\x1b[0m\n"
+        "    - \x1b[37;3mdef\x1b[0m\n"
+        "    - \x1b[37;3mghi\x1b[0m\n"
+        "    - \x1b[37;3mjkl\x1b[0m"
+    )
 
     # print("#3.4 - Case KVP-4:")
     obj = {'ab': Foo()}
-    result = ppc.format(obj, style=('red', 'italic'))
-    # print(f"{result}\n{[result]}")
-    assert result == ("\x1b[31;3mab:\x1b[0m \x1b[31;3mA Foo with:\x1b[0m\n"
-                      "\x1b[31;3m  - \x1b[33mk_1: v_1\x1b[0m\x1b[0m\n"
-                      "\x1b[31;3m  - \x1b[33mk_2: v_2\x1b[0m\x1b[0m")
+    result = ppc.format(obj, style=('white', 'italic'), key_style='red')
+    # print(f"\n{result}\n\n{[result]}")
+    assert result == (
+        "\x1b[31mab:\x1b[0m \x1b[37;3mA Foo with:\x1b[0m\n"
+        "\x1b[37;3m  \x1b[36m- k_1:\x1b[0m \x1b[36mv_1\x1b[0m\x1b[0m\n"
+        "\x1b[37;3m  \x1b[36m- k_2:\x1b[0m \x1b[36mv_2\x1b[0m\x1b[0m"
+    )
 
     # print("#3.5 - Case KVP-1, KVP-5:")
     obj = {'ab': 'xy', 'cd': Foo()}
-    result = ppc.format(obj, style=('red', 'italic'))
-    # print(f"{result}\n{[result]}")
-    assert result == ("- \x1b[31;3mab: xy\x1b[0m\n"
-                      "- \x1b[31;3mcd:\x1b[0m\n"
-                      "    \x1b[31;3mA Foo with:\x1b[0m\n"
-                      "    \x1b[31;3m- \x1b[33mk_1: v_1\x1b[0m\x1b[0m\n"
-                      "    \x1b[31;3m- \x1b[33mk_2: v_2\x1b[0m\x1b[0m")
+    result = ppc.format(obj, style=('white', 'italic'), key_style='red')
+    # print(f"\n{result}\n\n{[result]}")
+    assert result == (
+        "\x1b[31m- ab:\x1b[0m \x1b[37;3mxy\x1b[0m\n"
+        "\x1b[31m- cd:\x1b[0m\n"
+        "    \x1b[37;3mA Foo with:\x1b[0m\n"
+        "    \x1b[37;3m\x1b[36m- k_1:\x1b[0m \x1b[36mv_1\x1b[0m\x1b[0m\n"
+        "    \x1b[37;3m\x1b[36m- k_2:\x1b[0m \x1b[36mv_2\x1b[0m\x1b[0m"
+    )
 
 
 def test_ppc_style_4_extended_colors():